// frontend/src/app/models/gesture.ts
export enum Gesture {
    NONE = 'NONE',
    LEFT_HAND_UP = 'LEFT HAND UP!',
    RIGHT_HAND_UP = 'RIGHT HAND UP!',
    BOTH_HANDS_UP = 'BOTH HANDS UP!',
    HANDS_TOGETHER = 'HANDS TOGETHER!',
    LEFT_ARM_OUT = 'LEFT ARM OUT!',
    RIGHT_ARM_OUT = 'RIGHT ARM OUT!',
    BOTH_ARMS_OUT = 'BOTH ARMS OUT!'
}